+++ framework/py/flwr/supercore/superexec/executor/kubernetes_executor.py
@@ -20,7 +20,6 @@ import hashlib
 import importlib
 import json
 import re
-import threading
 import time
 from collections.abc import Callable, Mapping, Sequence
 from dataclasses import dataclass
@@ -39,9 +38,8 @@ from flwr.supercore.typing import JSONObject
 
 from .types import ExecutionSpec, LaunchResult
 from .warm_agentapp_executor import (
-    KubernetesWarmAgentAppDispatch,
-    WarmAgentAppUnavailable,
-    warm_agentapp_command,
+    WARM_EXECUTOR_CONSUMED_ANNOTATION,
+    WarmAgentAppPoolManager,
 )
 from .warm_executor import (
     WARM_EXECUTOR_MODULE,
@@ -57,7 +55,6 @@ from .warm_executor_pool import (
     WarmExecutorPoolKey,
     is_compatible_warm_executor,
     is_warm_executor,
-    is_warm_executor_ready,
     new_warm_executor_id,
 )
 
@@ -71,7 +68,7 @@ _COMPONENT_LABEL = "app.kubernetes.io/component"
 _TASK_TYPE_LABEL = "flower.ai/task-type"
 _RESOURCE_POOL_LABEL = "flower.ai/resource-pool"
 _WARM_EXECUTOR_OWNER_LABEL = "flower.ai/warm-executor-owner"
-_WARM_EXECUTOR_CONSUMED_ANNOTATION = "flower.ai/warm-executor-consumed"
+_WARM_EXECUTOR_CONSUMED_ANNOTATION = WARM_EXECUTOR_CONSUMED_ANNOTATION
 _EXECUTOR_OWNED_LABELS = frozenset(
     {
         _NAME_LABEL,
@@ -311,8 +308,8 @@ class KubernetesExecutorConfig:  # pylint: disable=too-many-instance-attributes
             )
 
 
-class _WarmExecutorPoolManager:  # pylint: disable=too-many-instance-attributes
-    """Own and dispatch a fixed set of compatible, one-task warm Pods."""
+class _WarmExecutorPoolManager(WarmAgentAppPoolManager):
+    """Wire the warm AgentApp pool lifecycle to Kubernetes executor helpers."""
 
     def __init__(
         self,
@@ -321,381 +318,16 @@ class _WarmExecutorPoolManager:  # pylint: disable=too-many-instance-attributes
         active_pod_count: Callable[[], int],
         exec_client: KubernetesClient | None = None,
     ) -> None:
-        self._client = client
-        self._exec_client = exec_client or client
-        self._config = config
-        self._active_pod_count = active_pod_count
-        self._pools = {pool.key: pool for pool in config.warm_executor_pools}
-        self._busy_pods: set[str] = set()
-        self._retiring_pods: set[str] = set()
-        self._closed = False
-        self._lock = threading.RLock()
-        self.ensure_capacity()
-
-    # pylint: disable-next=too-many-return-statements
-    def launch(
-        self, spec: ExecutionSpec, runtime_root_certificates: str | None
-    ) -> LaunchResult | None:
-        """Dispatch a compatible task to a ready Pod or use the cold fallback."""
-        pool = self._pool_for_spec(spec)
-        if pool is None:
-            return None
-
-        with self._lock:
-            if self._closed:
-                return None
-            try:
-                pod_name = self._take_ready_pod(pool.key)
-            except WarmAgentAppUnavailable:
-                return None
-            if pod_name is None:
-                self._ensure_pool_capacity(pool, reserved_pod_capacity=1)
-                return None
-            self._ensure_pool_capacity(pool)
-
-        try:
-            dispatch = self._open_dispatch(
-                pod_name=pod_name,
-                spec=spec,
-                runtime_root_certificates=runtime_root_certificates,
-            )
-        except WarmAgentAppUnavailable:
-            self._retire_unavailable_pod(pod_name)
-            return None
-
-        try:
-            dispatch.send_token(spec.token)
-        except Exception:  # pylint: disable=broad-exception-caught
-            self._retire_after_dispatch(pod_name, pool.key, dispatch)
-            return LaunchResult.unknown(
-                "Warm executor token delivery outcome was unknown."
-            )
-
-        try:
-            accepted = dispatch.wait_for_acceptance(_WARM_EXECUTOR_ACK_TIMEOUT_SECONDS)
-        except Exception:  # pylint: disable=broad-exception-caught
-            self._retire_after_dispatch(pod_name, pool.key, dispatch)
-            return LaunchResult.unknown(
-                "Warm executor token acknowledgement outcome was unknown."
-            )
-        self._retire_after_dispatch(pod_name, pool.key, dispatch)
-        if accepted:
-            return LaunchResult.accepted()
-        return LaunchResult.unknown(
-            "Warm executor did not acknowledge task token delivery."
+        super().__init__(
+            client,
+            config,
+            active_pod_count,
+            exec_client,
+            build_warm_executor_pod=_build_warm_executor_pod,
+            has_warm_executor_configuration=_has_warm_executor_configuration,
+            is_active_warm_executor=_is_active_warm_executor,
+            warm_executor_owner_label_selector=_warm_executor_owner_label_selector,
         )
-
-    def ensure_capacity(self, reserved_pod_capacity: int = 0) -> None:
-        """Create missing idle Pods for every configured compatible pool."""
-        with self._lock:
-            if self._closed:
-                return
-            self._reconcile_owned_pods()
-            for pool in self._pools.values():
-                self._ensure_pool_capacity(pool, reserved_pod_capacity)
-
-    def retry_retiring_pods(self) -> None:
-        """Retry Pod retirement without changing warm pool capacity."""
-        with self._lock:
-            if self._closed:
-                return
-            pods = self._owned_warm_pods()
-            if pods is not None:
-                self._retry_retiring_pods(pods)
-
-    def has_ready_pod(self, task_type: TaskType) -> bool:
-        """Return whether a matching warm Pod can take a task without new capacity."""
-        pool = self._pool_for_task(task_type)
-        if pool is None:
-            return False
-        with self._lock:
-            if self._closed:
-                return False
-            pods = self._owned_warm_pods()
-            if pods is None:
-                return False
-            return any(
-                (_object_name(pod) is not None)
-                and (_object_name(pod) not in self._busy_pods)
-                and not _is_consumed_warm_executor(pod)
-                and _has_warm_executor_configuration(pod, self._config)
-                and is_warm_executor_ready(pod, pool.key)
-                for pod in pods
-            )
-
-    def close(self) -> None:
-        """Stop dispatch and delete only idle Pods owned by this SuperExec."""
-        with self._lock:
-            if self._closed:
-                return
-            self._closed = True
-        pods = self._owned_warm_pods()
-        if pods is None:
-            return
-        for pod in pods:
-            pod_name = _object_name(pod)
-            if pod_name is not None and pod_name not in self._busy_pods:
-                self._delete_pod(pod_name)
-
-    def _pool_for_spec(self, spec: ExecutionSpec) -> WarmExecutorPoolConfig | None:
-        return self._pool_for_task(spec.task_type)
-
-    def _pool_for_task(self, task_type: TaskType) -> WarmExecutorPoolConfig | None:
-        for pool in self._pools.values():
-            if (
-                pool.key.task_type == task_type
-                and pool.key.runtime_image == self._config.image
-            ):
-                return pool
-        return None
-
-    def _take_ready_pod(self, key: WarmExecutorPoolKey) -> str | None:
-        pods = self._owned_warm_pods()
-        if pods is None:
-            raise WarmAgentAppUnavailable("Warm executor Pods could not be listed.")
-        for pod in pods:
-            pod_name = _object_name(pod)
-            if (
-                pod_name is not None
-                and pod_name not in self._busy_pods
-                and not _is_consumed_warm_executor(pod)
-                and _has_warm_executor_configuration(pod, self._config)
-                and is_warm_executor_ready(pod, key)
-            ):
-                try:
-                    self._mark_pod_consumed(pod_name)
-                except WarmAgentAppUnavailable:
-                    self._retire_unavailable_pod(pod_name)
-                    raise
-                self._busy_pods.add(pod_name)
-                return pod_name
-        return None
-
-    def _ensure_pool_capacity(
-        self, pool: WarmExecutorPoolConfig, reserved_pod_capacity: int = 0
-    ) -> None:
-        if self._retiring_pods:
-            return
-        pods = self._owned_warm_pods()
-        if pods is None:
-            return
-        compatible_count = sum(
-            1
-            for pod in pods
-            if _is_active_warm_executor(pod, pool.key, self._config)
-            and (
-                (
-                    _object_name(pod) not in self._busy_pods
-                    and not _is_consumed_warm_executor(pod)
-                )
-                or _object_name(pod) in self._retiring_pods
-            )
-        )
-        pods_to_create = max(pool.size - compatible_count, 0)
-        if self._config.active_pod_budget is not None:
-            try:
-                available_pod_capacity = (
-                    self._config.active_pod_budget
-                    - self._active_pod_count()
-                    - reserved_pod_capacity
-                )
-            except Exception:  # pylint: disable=broad-exception-caught
-                log(
-                    WARNING,
-                    "Warm executor capacity check failed; "
-                    "not creating replacement Pods.",
-                    exc_info=True,
-                )
-                return
-            pods_to_create = min(pods_to_create, max(available_pod_capacity, 0))
-        for _ in range(pods_to_create):
-            try:
-                pod = _build_warm_executor_pod(
-                    pool.key, self._config, new_warm_executor_id()
-                )
-                self._client.create_namespaced_pod(self._config.namespace, pod)
-            except Exception:  # pylint: disable=broad-exception-caught
-                log(WARNING, "Failed to create a warm TaskExecutor Pod.", exc_info=True)
-                return
-
-    def _reconcile_owned_pods(self) -> None:
-        """Delete owned Pods that are obsolete or exceed configured capacity."""
-        pods = self._owned_warm_pods()
-        if pods is None:
-            return
-        self._retry_retiring_pods(pods)
-
-        compatible_pods: dict[WarmExecutorPoolKey, list[object]] = {
-            pool.key: [] for pool in self._pools.values()
-        }
-        for pod in pods:
-            pod_name = _object_name(pod)
-            if (
-                pod_name is not None
-                and pod_name not in self._busy_pods
-                and _is_consumed_warm_executor(pod)
-            ):
-                self._retire_pod(pod_name)
-                continue
-            if pod_name in self._retiring_pods:
-                continue
-            pool = next(
-                (
-                    candidate
-                    for candidate in self._pools.values()
-                    if _is_active_warm_executor(pod, candidate.key, self._config)
-                ),
-                None,
-            )
-            if pool is None:
-                if pod_name is not None and pod_name not in self._busy_pods:
-                    self._retire_pod(pod_name)
-                continue
-            compatible_pods[pool.key].append(pod)
-
-        for pool in self._pools.values():
-            idle_pods = [
-                pod
-                for pod in compatible_pods[pool.key]
-                if _object_name(pod) not in self._busy_pods
-            ]
-            for pod in idle_pods[pool.size :]:
-                pod_name = _object_name(pod)
-                if pod_name is not None:
-                    self._retire_pod(pod_name)
-
-    def _retry_retiring_pods(self, pods: list[object]) -> None:
-        """Retry deletion of consumed Pods without making them dispatchable."""
-        pod_names = {
-            pod_name for pod in pods if (pod_name := _object_name(pod)) is not None
-        }
-        for pod_name in self._retiring_pods - pod_names:
-            self._release_pod(pod_name)
-        for pod in pods:
-            pod_name = _object_name(pod)
-            if pod_name is not None and pod_name in self._retiring_pods:
-                if self._delete_pod(pod_name):
-                    self._release_pod(pod_name)
-
-    def _owned_warm_pods(self) -> list[object] | None:
-        try:
-            pod_list = self._client.list_namespaced_pod(
-                self._config.namespace,
-                label_selector=_warm_executor_owner_label_selector(self._config),
-            )
-        except Exception:  # pylint: disable=broad-exception-caught
-            log(WARNING, "Failed to list warm TaskExecutor Pods.", exc_info=True)
-            return None
-        return _pod_items(pod_list)
-
-    def _open_dispatch(
-        self,
-        *,
-        pod_name: str,
-        spec: ExecutionSpec,
-        runtime_root_certificates: str | None,
-    ) -> KubernetesWarmAgentAppDispatch:
-        try:
-            stream = importlib.import_module("kubernetes.stream").stream
-            response = stream(
-                self._exec_client.connect_get_namespaced_pod_exec,
-                pod_name,
-                self._config.namespace,
-                command=warm_agentapp_command(spec, runtime_root_certificates),
-                stderr=True,
-                stdin=True,
-                stdout=True,
-                tty=False,
-                _preload_content=False,
-            )
-        except Exception as err:  # pylint: disable=broad-exception-caught
-            raise WarmAgentAppUnavailable(
-                "Warm TaskExecutor Pod is unavailable for dispatch."
-            ) from err
-        return KubernetesWarmAgentAppDispatch(response)
-
-    def _mark_pod_consumed(self, pod_name: str) -> None:
-        try:
-            self._client.patch_namespaced_pod(
-                name=pod_name,
-                namespace=self._config.namespace,
-                body={
-                    "metadata": {
-                        "annotations": {
-                            _WARM_EXECUTOR_CONSUMED_ANNOTATION: "true",
-                        }
-                    }
-                },
-            )
-        except Exception as err:  # pylint: disable=broad-exception-caught
-            raise WarmAgentAppUnavailable(
-                "Failed to persist warm TaskExecutor Pod consumption."
-            ) from err
-
-    def _retire_after_dispatch(
-        self,
-        pod_name: str,
-        key: WarmExecutorPoolKey,
-        dispatch: KubernetesWarmAgentAppDispatch,
-    ) -> None:
-        cleanup_thread = threading.Thread(
-            target=self._wait_for_task_and_replace,
-            args=(pod_name, key, dispatch),
-            daemon=True,
-        )
-        try:
-            cleanup_thread.start()
-        except RuntimeError:
-            self._wait_for_task_and_replace(pod_name, key, dispatch)
-
-    def _wait_for_task_and_replace(
-        self,
-        pod_name: str,
-        key: WarmExecutorPoolKey,
-        dispatch: KubernetesWarmAgentAppDispatch,
-    ) -> None:
-        try:
-            dispatch.wait_for_close()
-        finally:
-            try:
-                dispatch.close()
-            except Exception:  # pylint: disable=broad-exception-caught
-                log(WARNING, "Failed to close warm TaskExecutor exec stream.")
-            if self._retire_pod(pod_name):
-                with self._lock:
-                    if not self._closed:
-                        self._ensure_pool_capacity(self._pools[key])
-
-    def _release_pod(self, pod_name: str) -> None:
-        with self._lock:
-            self._busy_pods.discard(pod_name)
-            self._retiring_pods.discard(pod_name)
-
-    def _retire_unavailable_pod(self, pod_name: str) -> None:
-        """Delete a Pod that failed before task authority was delivered."""
-        self._retire_pod(pod_name)
-
-    def _retire_pod(self, pod_name: str) -> bool:
-        with self._lock:
-            self._retiring_pods.add(pod_name)
-        if self._delete_pod(pod_name):
-            self._release_pod(pod_name)
-            return True
-        return False
-
-    def _delete_pod(self, pod_name: str) -> bool:
-        try:
-            self._client.delete_namespaced_pod(
-                name=pod_name,
-                namespace=self._config.namespace,
-                grace_period_seconds=0,
-            )
-        except Exception as exc:  # pylint: disable=broad-exception-caught
-            if _exception_status(exc) == 404:
-                return True
-            log(WARNING, "Failed to delete warm TaskExecutor Pod %s.", pod_name)
-            return False
-        return True
 
 
 class KubernetesExecutor:
